Re-organize subroutines and functions, moving non-communication routines from comm.f90, comm_mpi.f90 and comm_ser.f90 to utils.f90 by using pre- processor variable MPI. Change all real variables (e.g. areas) to real(8). Minor cleanup.

+++ src/multi_level.f90
@@ -480,7 +480,7 @@ contains
       integer, dimension(2,N_BDRY+1) :: dims
 
       type(Domain)           :: dom
-      real(4)                :: wgt
+      real(8)                :: wgt
       integer, dimension(20) :: id
 
       call get_indices (dom, i+1, j, RT, offs, dims, id)
