+++ src/patch.f90
@@ -24,15 +24,15 @@ module patch_mod
 
   type Overl_Area
      real(8), dimension(4) :: a
-     real(4), dimension(2) :: split  
+     real(8), dimension(2) :: split  
   end type Overl_Area
 
   type Iu_Wgt
-     real(4), dimension(9) :: enc  
+     real(8), dimension(9) :: enc  
   end type Iu_Wgt
 
   type RF_Wgt
-     real(4), dimension(3) :: enc
+     real(8), dimension(3) :: enc
   end type RF_Wgt
 contains
   subroutine init_patch_mod
@@ -83,7 +83,7 @@ contains
   subroutine init_RF_Wgt (self, wgt)
     implicit none
     type(RF_Wgt)          :: self
-    real(4), dimension(3) :: wgt
+    real(8), dimension(3) :: wgt
 
     self%enc = wgt
   end subroutine init_RF_Wgt
